improve: 【快速分发文件】插件v1.0，不校验源文件是否在当前业务下，允许跨业务分发
